Add the required styling
Two ids were created for the divs, the first with blue text and font size of 36px the second with a light green background and font size of 24px and bold.
Two classes were created; the first applied to all the paragraphs giving them a light pink background with and a list of fonts containing Verdana and DejaVu Sans with sans-serif as a fallback.
The second class applied only to the second paragraph and increased the size of the text.